Remove load Pkl modules directly from VaultClient object
